update on assign fworker name
