+++ dfttk/scripts/assign_fworker_name.py
@@ -1,0 +1,78 @@
+
+
+from atomate.utils.utils import get_meta_from_structure, get_fws_and_tasks
+def set_queue_options(
+    original_wf,
+    walltime=None,
+    time_min=None,
+    qos=None,
+    pmem=None,
+    fw_name_constraint=None,
+    task_name_constraint=None,
+):
+    """
+    Modify queue submission parameters of Fireworks in a Workflow.
+    This powerup overrides paramters in the qadapter file by setting values in
+    the 'queueadapter' key of a Firework spec. For example, the walltime
+    requested from a queue can be modified on a per-workflow basis.
+    Args:
+        original_wf (Workflow):
+        walltime (str): Total walltime to request for the job in HH:MM:SS
+            format e.g., "00:10:00" for 10 minutes.
+        time_min (str): Minimum walltime to request in HH:MM:SS format.
+            Specifying both `walltime` and `time_min` can improve throughput on
+            some queues.
+        qos (str): QoS level to request. Typical examples include "regular",
+            "flex", and "scavenger". For Cori KNL "flex" QoS, it is necessary
+            to specify a `time_min` of no more than 2 hours.
+        fw_name_constraint (str): name of the Fireworks to be tagged (all if
+            None is passed)
+        task_name_constraint (str): name of the Firetasks to be tagged (e.g.
+            None or 'RunVasp')
+    Returns:
+        Workflow: workflow with modified queue options
+    """
+    qsettings = {}
+    if walltime:
+        qsettings.update({"walltime": walltime})
+    if time_min:
+        qsettings.update({"time_min": time_min})
+    if qos:
+        qsettings.update({"qos": qos})
+    if pmem:
+        qsettings.update({"pmem": pmem})
+
+    idx_list = get_fws_and_tasks(
+        original_wf,
+        fw_name_constraint=fw_name_constraint,
+        task_name_constraint=task_name_constraint,
+    )
+
+    for idx_fw, idx_t in idx_list:
+        original_wf.fws[idx_fw].spec.update({"_queueadapter": qsettings})
+
+    return original_wf
+
+
+import atomate.vasp.powerups as powerups
+def Customizing_Workflows(original_wf, user_settings={}):
+    """
+    if os.path.exists('SETTINGS.yaml'):
+        user_settings= loadfn('SETTINGS.yaml')
+    else:
+        user_settings={}
+    #ymal dict, see https://atomate.org/customizing_workflows.html
+    """
+    powerups_options = user_settings.get('powerups', {})
+    if len(powerups_options) == 0:
+        if 'user_incar_settings' in user_settings:
+            powerups_options = user_settings['user_incar_settings'].get('powerups', {})
+    if 'set_execution_options' in powerups_options:
+        execution_options = powerups_options['set_execution_options']
+        original_wf = powerups.set_execution_options(original_wf, 
+            fworker_name=execution_options.get("fworker_name", None))
+    if 'set_queue_options' in powerups_options:
+        queue_options = powerups_options['set_queue_options']
+        #print(queue_options,"eeeeeeeeeeee")
+        original_wf = set_queue_options(original_wf, pmem=queue_options.get("pmem", None))
+    return original_wf
