+++ dfttk/scripts/run_dfttk.py
@@ -7,6 +7,7 @@ from dfttk.wflows import get_wf_gibbs, get_wf_EV_bjb, get_wf_gibbs_robust, get_w
 from dfttk.utils import recursive_glob
 from dfttk.structure_builders.parse_anrl_prototype import multi_replace
 from dfttk.scripts.querydb import get_eq_structure_by_metadata
+from dfttk.scripts.assign_fworker_name import Customizing_Workflows
 import dfttk.scripts.querydb as querydb
 from fireworks.fw_config import config_to_dict
 from monty.serialization import loadfn, dumpfn
@@ -174,6 +175,10 @@ def get_wf_single(structure, WORKFLOW="get_wf_gibbs", settings={}):
     #override_default_vasp_params = {'user_incar_settings': {}, 'user_kpoints_settings': {}, 'user_potcar_functional': str}
     #If some value in 'user_incar_settings' is set to None, it will use vasp's default value
     override_default_vasp_params = settings.get('override_default_vasp_params', {})
+    #check if fworker_name is assigned
+    powerups = settings.get('powerups', {})
+    override_default_vasp_params['user_incar_settings'].update({'powerups':powerups})
+
     #dict, dict of class ModifyIncar with keywords in Workflow name. e.g.
     """
     modify_incar_params = { 'Full relax': {'incar_update': {"LAECHG":False,"LCHARG":False,"LWAVE":False}},
@@ -269,78 +274,6 @@ def get_wf_single(structure, WORKFLOW="get_wf_gibbs", settings={}):
     return wf
 
 
-from atomate.utils.utils import get_meta_from_structure, get_fws_and_tasks
-def set_queue_options(
-    original_wf,
-    walltime=None,
-    time_min=None,
-    qos=None,
-    pmem=None,
-    fw_name_constraint=None,
-    task_name_constraint=None,
-):
-    """
-    Modify queue submission parameters of Fireworks in a Workflow.
-    This powerup overrides paramters in the qadapter file by setting values in
-    the 'queueadapter' key of a Firework spec. For example, the walltime
-    requested from a queue can be modified on a per-workflow basis.
-    Args:
-        original_wf (Workflow):
-        walltime (str): Total walltime to request for the job in HH:MM:SS
-            format e.g., "00:10:00" for 10 minutes.
-        time_min (str): Minimum walltime to request in HH:MM:SS format.
-            Specifying both `walltime` and `time_min` can improve throughput on
-            some queues.
-        qos (str): QoS level to request. Typical examples include "regular",
-            "flex", and "scavenger". For Cori KNL "flex" QoS, it is necessary
-            to specify a `time_min` of no more than 2 hours.
-        fw_name_constraint (str): name of the Fireworks to be tagged (all if
-            None is passed)
-        task_name_constraint (str): name of the Firetasks to be tagged (e.g.
-            None or 'RunVasp')
-    Returns:
-        Workflow: workflow with modified queue options
-    """
-    qsettings = {}
-    if walltime:
-        qsettings.update({"walltime": walltime})
-    if time_min:
-        qsettings.update({"time_min": time_min})
-    if qos:
-        qsettings.update({"qos": qos})
-    if pmem:
-        qsettings.update({"pmem": pmem})
-
-    idx_list = get_fws_and_tasks(
-        original_wf,
-        fw_name_constraint=fw_name_constraint,
-        task_name_constraint=task_name_constraint,
-    )
-
-    for idx_fw, idx_t in idx_list:
-        original_wf.fws[idx_fw].spec.update({"_queueadapter": qsettings})
-
-    return original_wf
-
-
-import atomate.vasp.powerups as powerups
-def Customizing_Workflows(original_wf):
-    if os.path.exists('SETTINGS.yaml'):
-        user_settings= loadfn('SETTINGS.yaml')
-    else:
-        user_settings={}
-    #ymal dict, see https://atomate.org/customizing_workflows.html
-    powerups_options = user_settings.get('powerups', {})
-    if 'set_execution_options' in powerups_options:
-        execution_options = powerups_options['set_execution_options']
-        original_wf = powerups.set_execution_options(original_wf, 
-            fworker_name=execution_options.get("fworker_name", None))
-    if 'set_queue_options' in powerups_options:
-        queue_options = powerups_options['set_queue_options']
-        #print(queue_options,"eeeeeeeeeeee")
-        original_wf = set_queue_options(original_wf, pmem=queue_options.get("pmem", None))
-    return original_wf
-
 def run(args):
     """
     Run dfttk
@@ -409,6 +342,7 @@ def run(args):
                 user_settings.update({"phonon_supercell_matrix": "atoms"})
 
             wf = get_wf_single(structure, WORKFLOW=WORKFLOW, settings=user_settings)
+            wf = Customizing_Workflows(wf,user_settings=user_settings)
             if isinstance(wf, list):
                 wfs = wfs + wf
             else:
@@ -457,7 +391,7 @@ def run(args):
                         user_settings.update({"phonon_supercell_matrix": "atoms"})
 
                     wf = get_wf_single(structure, WORKFLOW=WORKFLOW, settings=user_settings)
-
+                    wf = Customizing_Workflows(wf,user_settings=user_settings)
                     metadatas[STR_FILE] = wf.as_dict()["metadata"]
                     wfs.append(wf)
 
@@ -468,11 +402,13 @@ def run(args):
         #Write Out the metadata for POST and continue purpose
         dumpfn(metadatas, "METADATAS.yaml")
 
+    """
     _fws = []
     for wflow in wfs:
-        revised_wflow = Customizing_Workflows(wflow)
+        revised_wflow = Customizing_Workflows(wflow,user_settings={})
         _fws.append(revised_wflow)
     fws = _fws
+    """
 
     if LAUNCH:
         from fireworks import LaunchPad
